add nrutil.c downloaded (free) from numerical book recipies website
